+++ tests/test_tileset.py
@@ -1,0 +1,8 @@
+import tcod
+
+
+def test_proc_block_elements() -> None:
+    tileset = tcod.tileset.Tileset(8, 8)
+    tcod.tileset.procedural_block_elements(tileset)
+    tileset = tcod.tileset.Tileset(0, 0)
+    tcod.tileset.procedural_block_elements(tileset)